requirements.txt
aniso8601==9.0.1
astroid==2.14.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
dill==0.3.6
flake8==6.0.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
geographiclib==2.0
geopy==2.3.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jplephem==2.18
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pygame==2.1.3.dev8
PyJWT==2.6.0
pylint==2.16.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
sgp4==2.21
six==1.16.0
skyfield==1.45
SQLAlchemy==2.0.9
tomlkit==0.11.6
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.14.1
